// src/app/models/company.ts
export interface Company {
    id: number;
    companyName: String;
    companyPhoneNumber: String;
    companyFaxNumber: String;
    companyEmail: String;
    companyAddress: String;
    commercialCompanyName: String;
    commercialNumber: String;
    commercialDocument: String;
    commercialIssuedPlace: String;
    taxCompanyName: String;
    taxNumber: String;
    taxDocument: String;
    taxIssuedPlace: String;
    companyRepresenters: any[];
}