// src/app/models/representer.ts
export interface Representer {
    id: number;
    representerFullName: String;
    representerOfficeNumber: String;
    representerEmail: String;
    representerPIOID: String;
    representerFaxNumber: String;
    representerNumber: String;
    isOnwer: boolean;
    companyId: number;
    jobTypeId: number;
    jobTypeName: string;
}